refactor: ChessGame -> ScoreCalculator, PieceMover
