+++ src/main/java/softeer2nd/chess/PieceMover.java
@@ -7,35 +7,25 @@ import softeer2nd.chess.domain.pieces.enums.PieceColor;
 import softeer2nd.chess.domain.pieces.enums.PieceDirection;
 
 import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
-import static softeer2nd.chess.domain.Board.BOARD_LENGTH;
-
-public class ChessGame {
-    private final Board board;
-
-    public ChessGame(Board board) {
-        this.board = board;
-    }
-
-    public void movePiece(Position sourcePosition, Position targetPosition, PieceColor color) {
+public class PieceMover {
+    public void movePiece(Board board, Position sourcePosition, Position targetPosition, PieceColor color) {
         Piece piece = board.findPieceByPosition(sourcePosition);
         verifyMatchColor(piece, color);
 
         List<Position> positionsOnRoute = piece.getPositionsOnRoute(sourcePosition, targetPosition);
         verifyRouteExist(positionsOnRoute);
-        verifyNoPieceOnRoute(positionsOnRoute, color);
+        verifyNoPieceOnRoute(board, positionsOnRoute, color);
 
         if (piece.isPawn()) {
-            verifyPawnMove(sourcePosition, targetPosition);
+            verifyPawnMove(board, sourcePosition, targetPosition);
         }
 
         board.removePiece(sourcePosition);
         board.addPiece(targetPosition, piece);
     }
 
-    private void verifyPawnMove(Position sourcePosition, Position targetPosition) {
+    private void verifyPawnMove(Board board, Position sourcePosition, Position targetPosition) {
         int diffRank = targetPosition.getRankIndex() - sourcePosition.getRankIndex();
         int diffFile = targetPosition.getFileIndex() - sourcePosition.getFileIndex();
 
@@ -53,7 +43,7 @@ public class ChessGame {
         }
     }
 
-    private void verifyNoPieceOnRoute(List<Position> positionsOnRoute, PieceColor color) {
+    private void verifyNoPieceOnRoute(Board board, List<Position> positionsOnRoute, PieceColor color) {
         Position destination = positionsOnRoute.remove(positionsOnRoute.size() - 1);
         for (Position position : positionsOnRoute) {
             Piece pieceOnRoute = board.findPieceByPosition(position);
@@ -81,45 +71,5 @@ public class ChessGame {
             throw new RuntimeException("해당 위치에 기물이 없습니다.");
         }
         throw new RuntimeException(color.name() + "차례 입니다.");
-    }
-
-    public double calculatePoint(PieceColor color) {
-        return IntStream.range(0, BOARD_LENGTH)
-                .mapToDouble(file -> calculateFilePoint(file, color))
-                .sum();
-    }
-
-    private double calculateFilePoint(int fileIndex, PieceColor color) {
-        List<Piece> piecesInFile = findPiecesInFileByColor(fileIndex, color);
-
-        double point = calPoint(piecesInFile);
-        int numOfPawns = getNumOfPawns(piecesInFile);
-        return point - calPenaltyPoint(numOfPawns);
-    }
-
-    private List<Piece> findPiecesInFileByColor(int fileIndex, PieceColor color) {
-        return board.findPiecesInFile(fileIndex)
-                .stream()
-                .filter(piece -> piece.matchesColor(color))
-                .collect(Collectors.toList());
-    }
-
-    private int getNumOfPawns(List<Piece> pieces) {
-        return (int) pieces.stream()
-                .filter(Piece::isPawn)
-                .count();
-    }
-
-    private double calPoint(List<Piece> pieces) {
-        return pieces.stream()
-                .mapToDouble(Piece::getDefaultPoint)
-                .sum();
-    }
-
-    private double calPenaltyPoint(int numOfPawn) {
-        if (numOfPawn >= 2) {
-            return numOfPawn * 0.5;
-        }
-        return 0;
     }
 }
